Extracted create-post-component and included it on landing-page.
