+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/create-need/create-need.html
@@ -2,92 +2,6 @@
     <won-topnav></won-topnav>
     <won-create-need-title-bar></won-create-need-title-bar>
 </header>
-<main class="create-post">
-    <div class="cp__inner">
-
-        <won-posttype-select
-                options="::self.postTypeTexts"
-                on-select="::self.selectType(idx)"
-                on-unselect="::self.unselectType()">
-        </won-posttype-select>
-
-        <div class="cp__mandatory-rest" ng-show="self.draft.type">
-            <won-image-dropzone
-                    on-image-picked="::self.pickImage(image)">
-            </won-image-dropzone>
-
-            <need-textfield on-draft-change="::self.setDraft(draft)"></need-textfield>
-        </div>
-
-        <hr ng-show="self.draft.tags && self.draft.tags.length > 0"/>
-        <div class="cp__tags" ng-show="self.draft.tags && self.draft.tags.length > 0">
-            <div class="cp__header tags">Tags</div>
-            <span class="cp__tags__tag" ng-repeat="tag in self.draft.tags">{{tag}}</span>
-        </div>
-
-
-        <hr ng-show="self.isValid()"/>
-        <div class="cp__location" ng-show="self.isValid()">
-            <div class="cp__header location">Location</div>
-            <won-location-picker on-draft-change="::self.setDraft(draft)" location-is-saved="::self.locationIsSaved()"></won-location-picker>
-        </div>
-
-        <!--
-        <div>
-            <p>{First time} Roughly, what are you looking for? This will shop up as title. </p>
-            <a href="javascript:void(0)">Don't show again</a>
-        </div>
-
-        <div class="cp__hr"></div>
-
-        <div class="cp__more-detail">
-            <img src="generated/icon-sprite.svg#ico36_plus_circle" class="cp__icon">
-            <a href="javascript:void(0)">Add more Detail</a>
-        </div>
-
-        <won-labelled-hr label="::undefined"></won-labelled-hr>
-        -->
-        <!--<div class="cp__more-detail--folded-out" ng-show="self.isValid()">
-            <div class="cp__more-detail__item">
-                <img src="generated/icon-sprite.svg#ico36_plus_circle_hi" class="cp__icon">
-                <a href="javascript:void(0)">Add more Detail</a>
-            </div>
-            <ul>
-                <li>
-                    <img src="generated/icon-sprite.svg#ico36_media_circle" class="cp__icon">
-                    <a href="javascript:void(0)">Image or Media</a>
-                </li>
-                <li>
-                    <img src="generated/icon-sprite.svg#ico36_location_circle" class="cp__icon">
-                    <a href="javascript:void(0)">Address or Location</a>
-                </li>
-                <li>
-                    <img src="generated/icon-sprite.svg#ico36_area_circle" class="cp__icon">
-                    <a href="javascript:void(0)">Area or Radius</a>
-                </li>
-                <li>
-                    <img src="generated/icon-sprite.svg#ico36_time_circle" class="cp__icon">
-                    <a href="javascript:void(0)">Deadline or Timeframe</a>
-                </li>
-                <li>
-                    <img src="generated/icon-sprite.svg#ico36_notification_circle" class="cp__icon">
-                    <a href="javascript:void(0)">Manage Notifications</a>
-                </li>
-            </ul>
-        </div>-->
-
-        <won-labelled-hr label="::'or'" ng-show="self.isValid()"></won-labelled-hr>
-
-        <button type="submit" class="won-button--filled red"
-                ng-show="self.isValid()"
-                ng-click="::self.publish()">
-            <span ng-show="!self.pendingPublishing">
-                Publish
-            </span>
-            <span ng-show="self.pendingPublishing">
-                Publishing&nbsp;&hellip;
-            </span>
-        </button>
-
-    </div>
+<main>
+    <won-create-post></won-create-post>
 </main>
