+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/landingpage/landingpage.html
@@ -3,6 +3,9 @@
 </header>
 <main class="landingpage" id="allSections">
     <a class="curtain" ng-show="self.focusSignup" ng-click="self.router__stateGoCurrent({focusSignup: false})"></a>
+
+    <won-create-post></won-create-post>
+
     <div class="welcome">
         <div class="title">Welcome to the Web of Needs</div>
         <won-flex-grid items="::self.peopleGrid"></won-flex-grid>
